+++ client/src/components/Icons.jsx
@@ -1,5 +1,5 @@
-import { ReactComponent as Search } from './assets/search.svg';
-import { ReactComponent as User } from './assets/user.svg';
+import { ReactComponent as Search } from '../assets/search.svg';
+import { ReactComponent as User } from '../assets/user.svg';
 
 const validateClassName = className => {
   return typeof className === 'string' && className.length > 0 ? { className } : '';
